requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
bootstrap4==0.1.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap3==2024.1
crispy-bootstrap4==2024.1
cron-descriptor==1.4.5
cryptography==43.0.3
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==4.2.16
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.52.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-mathfilters==1.0.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-auth==2.35.0
google-auth-oauthlib==1.2.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
python-barcode==0.15.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
redis==5.1.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
stripe==11.1.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
